# latihan1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Hactiv8 Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/api/v1",
    "paths": {
        "/Ping": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.PingSwag"
                        }
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "description": "get all todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Get All TODOS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetTodosSuccessSwag"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.GetTodosFailureSwag"
                        }
                    }
                }
            },
            "post": {
                "description": "create todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Create TODO",
                "parameters": [
                    {
                        "description": "CreateToDo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.CreateTodoSuccessSwag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/views.CreateTodoFailureSwag"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "Get Todo By Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Get Todo By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetTodoByIdSuccessSwag"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.GetTodosFailureSwag"
                        }
                    }
                }
            },
            "put": {
                "description": "Update TODO By Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Update TODO By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateTodo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.UpdateTodoByIdSuccessSwag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/views.UpdateTodoFailureSwag"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete TODO by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Delete TODO by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.DeleteTodoByIdSuccessSwag"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.DeleteTodoFailureSwag"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.PingSwag": {
            "type": "object",
            "properties": {
                "health": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "request.CreateTodo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "views.CreateTodoFailureSwag": {
            "type": "object",
            "properties": {
                "additional_info": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "error": "Title harus diisi"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "message": {
                    "type": "string",
                    "example": "Create Todo Fail"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "views.CreateTodoPayload": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "views.CreateTodoSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Create Todo Success"
                },
                "payload": {
                    "$ref": "#/definitions/views.CreateTodoPayload"
                },
                "status": {
                    "type": "integer",
                    "example": 201
                }
            }
        },
        "views.DeleteTodoByIdSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Get Todo By ID Success"
                },
                "payload": {
                    "$ref": "#/definitions/views.GetTodoPayload"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "views.DeleteTodoFailureSwag": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Data Not Found"
                },
                "message": {
                    "type": "string",
                    "example": "Delete Todo Fail"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "views.GetTodoByIdSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Get Todo By ID Success"
                },
                "payload": {
                    "$ref": "#/definitions/views.GetTodoPayload"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "views.GetTodoPayload": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "views.GetTodosFailureSwag": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Data Not Found"
                },
                "message": {
                    "type": "string",
                    "example": "Data Is Empty"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "views.GetTodosSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Get All Todos Success"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.GetTodoPayload"
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "views.UpdateTodoByIdSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Get Todo By ID Success"
                },
                "payload": {
                    "$ref": "#/definitions/views.GetTodoPayload"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "views.UpdateTodoFailureSwag": {
            "type": "object",
            "properties": {
                "additional_info": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "error": "Title harus diisi"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "message": {
                    "type": "string",
                    "example": "Update Todo Fail"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        }
    }
}